adding back bluestein parts bit by bit
