+++ CMakeLists.txt
@@ -48,7 +48,8 @@ option(PORTFFT_CLANG_TIDY "Enable clang-tidy checks on portFFT source when build
 option(PORTFFT_CLANG_TIDY_AUTOFIX "Attempt to fix defects found by clang-tidy" OFF)
 option(PORTFFT_LOG_DUMPS "Whether to enable logging of data dumps" OFF)
 option(PORTFFT_LOG_TRANSFERS "Whether to enable logging of memory transfers" OFF)
-option(PORTFFT_LOG_TRACE "Whether to enable tracing of function calls" OFF)
+option(PORTFFT_LOG_TRACES "Whether to enable tracing of function calls" OFF)
+option(PORTFFT_LOG_WARNINGS "Whether to enable logging of warnings" ON)
 set(PORTFFT_REGISTERS_PER_WI 128 CACHE STRING "How many 32b registers can be allocated per work item on the target device")
 set(PORTFFT_SUBGROUP_SIZES 32 CACHE STRING "Comma separated list of subgroup sizes to compile for. The first size supported by the device will be used.")
 set(PORTFFT_VEC_LOAD_BYTES 16 CACHE STRING "Number of consecutive bytes each work item should load at once.")
@@ -92,21 +93,24 @@ if(${PORTFFT_ENABLE_OOP_BUILDS})
   target_compile_definitions(portfft INTERFACE PORTFFT_ENABLE_OOP_BUILDS)
 endif()
 
-set(PORTFFT_LOG OFF)
+set(PORTFFT_KERNEL_LOG OFF)
 if(${PORTFFT_LOG_DUMPS})
   target_compile_definitions(portfft INTERFACE PORTFFT_LOG_DUMPS)
-  set(PORTFFT_LOG ON)
+  set(PORTFFT_KERNEL_LOG ON)
 endif()
 if(${PORTFFT_LOG_TRANSFERS})
   target_compile_definitions(portfft INTERFACE PORTFFT_LOG_TRANSFERS)
-  set(PORTFFT_LOG ON)
+  set(PORTFFT_KERNEL_LOG ON)
 endif()
-if(${PORTFFT_LOG_TRACE})
-  target_compile_definitions(portfft INTERFACE PORTFFT_LOG_TRACE)
-  set(PORTFFT_LOG ON)
+if(${PORTFFT_LOG_TRACES})
+  target_compile_definitions(portfft INTERFACE PORTFFT_LOG_TRACES)
+  set(PORTFFT_KERNEL_LOG ON)
 endif()
-if(${PORTFFT_LOG})
-  target_compile_definitions(portfft INTERFACE PORTFFT_LOG)
+if(${PORTFFT_LOG_WARNINGS})
+  target_compile_definitions(portfft INTERFACE PORTFFT_LOG_WARNINGS)
+endif()
+if(${PORTFFT_KERNEL_LOG})
+  target_compile_definitions(portfft INTERFACE PORTFFT_KERNEL_LOG)
 endif()
 
 target_compile_options(portfft INTERFACE -fgpu-inline-threshold=1000000)
