+++ src/portfft/common/host_fft.hpp
@@ -18,6 +18,7 @@
  *
  **************************************************************************/
 
+
 #ifndef PORTFFT_COMMON_HOST_FFT_HPP
 #define PORTFFT_COMMON_HOST_FFT_HPP
