+++ src/portfft/specialization_constant.hpp
@@ -44,9 +44,8 @@ constexpr static sycl::specialization_id<level> GlobalSubImplSpecConst{};
 constexpr static sycl::specialization_id<Idx> GlobalSpecConstLevelNum{};
 constexpr static sycl::specialization_id<Idx> GlobalSpecConstNumFactors{};
 
-constexpr static sycl::specialization_id<bool> SpecConstIncrementModifierPointer{};
 // Specialization constants used for IFFT, when expressed as a IFFT=(conjugate(FFT(conjugate(input))))
-constexpr static sycl::specialization_id<detail::complex_conjugate> SpecTakeConjugateOnLoad{};
+constexpr static sycl::specialization_id<detail::complex_conjugate> SpecConstConjugateOnLoad{};
 constexpr static sycl::specialization_id<detail::complex_conjugate> SpecConstConjugateOnStore{};
 
 constexpr static sycl::specialization_id<float> SpecConstScaleFactorFloat{};
