+++ test/unit_test/instantiate_fft_tests.hpp
@@ -141,12 +141,6 @@ INSTANTIATE_TEST_SUITE_P(GlobalTest, FFTTest,
                              ::testing::Values(sizes_t{32768}, sizes_t{65536}, sizes_t{131072}))),
                          test_params_print());
 
-INSTANTIATE_TEST_SUITE_P(PrimeSizedTest, FFTTest,
-                         ::testing::ConvertGenerator<basic_param_tuple>(::testing::Combine(
-                             all_valid_global_placement_layouts, fwd_only, interleaved_storage, ::testing::Values(1, 3),
-                             ::testing::Values(sizes_t{211}, sizes_t{523}, sizes_t{65537}))),
-                         test_params_print());
-
 INSTANTIATE_TEST_SUITE_P(WorkgroupOrGlobalRegressionTest, FFTTest,
                          ::testing::ConvertGenerator<basic_param_tuple>(
                              ::testing::Combine(ip_packed_layout, fwd_only, interleaved_storage, ::testing::Values(3),
